requirements.txt
streamlit
pandas
odfpy
openpyxl
schedule>=1.2.0 
tabula-py
pytz
